order microservice with connection error

+++ internal/models/product.go
@@ -8,6 +8,6 @@ type Product struct {
 	Description string    `json:"description,omitempty"`
 	Price       int64     `json:"price"`
 	ImgSrc      string    `json:"img"`
-	Rating      float64   `json:"rating"`
+	Rating      float32   `json:"rating"`
 	Category
 }
